ref:Moved user fixture to global conftest for use across app tests
