+++ dweller/accounts/tests/conftest.py
@@ -1,15 +1,6 @@
 import pytest
 from datetime import date, timedelta
-
-@pytest.fixture
-def normal_user():  
-    return {
-        'email': "normal@user.com",
-        'password': "foo",
-        'first_name': "John",
-        'last_name': "Doe",
-        'dob': "1990-01-01"
-    }
+from tests.conftest_global import normal_user
 
 @pytest.fixture
 def valid_form_data():
