+++ dweller/tests/conftest_global.py
@@ -1,0 +1,11 @@
+import pytest
+
+@pytest.fixture
+def normal_user():  
+    return {
+        'email': "normal@user.com",
+        'password': "foo",
+        'first_name': "John",
+        'last_name': "Doe",
+        'dob': "1990-01-01"
+    }
